Add a  bunch of check for when deleting, editing video, playlist. Fix adset check bugs. Add route for displaying containing ads

+++ controllers/adSet.js
@@ -1,6 +1,7 @@
 const adSetService = require("./../services/adSet");
 const config = require("./../config/config");
 const adSetSupport = require("./../utils/getAdSetStatus");
+
 async function insert(req, res) {
   try {
     const { name, daysOfWeek, hoursOfDay } = req.body;
@@ -73,16 +74,16 @@ async function updateById(req, res) {
   try {
     const { id } = req.params;
     const { daysOfWeek, hoursOfDay } = req.body;
-    let document = adSetService.getById(id);
+    let document = await adSetService.getById(id);
     if (document["adManagerId"].toString() != req.userId) {
       return res
         .status(config.status_code.FORBIDEN)
         .send({ message: "wrong user" });
     }
-    let { result, adOfferDoc } = adSetSupport.getAdSetStatus(id);
-    if (result == false) {
+    let { result, adOfferDoc } = await adSetSupport.getAdSetStatus(id);
+    if (result === false) {
       return res.status(config.status_code.FORBIDEN).send({
-        message: "adSet is using by some adOffer or playlist in adOffer",
+        message: "Some active ads are using this setting",
         adOffers: adOfferDoc,
       });
     }
@@ -90,7 +91,7 @@ async function updateById(req, res) {
       daysOfWeek,
       hoursOfDay,
     });
-    document = adSetService.getById(id);
+    document = await adSetService.getById(id);
     return res.status(config.status_code.OK).send({ adset: document });
   } catch (error) {
     console.log(error);
@@ -101,11 +102,18 @@ async function updateById(req, res) {
 async function deleteById(req, res) {
   try {
     const { id } = req.params;
-    let document = adSetService.getById(id);
+    let document = await adSetService.getById(id);
     if (document["adManagerId"].toString() != req.userId) {
       return res
         .status(config.status_code.FORBIDEN)
         .send({ message: "wrong user" });
+    }
+    let { result, adOfferDoc } = await adSetSupport.getAdSetStatus(id);
+    if (result === false) {
+      return res.status(config.status_code.FORBIDEN).send({
+        message: "Some active ads are using this setting",
+        adOffers: adOfferDoc,
+      });
     }
     await adSetService.deleteById(id);
     return res.status(config.status_code.OK).send({ adset: true });
